Use the html body in the example template
The body_html returned by the GitHub API is much better suited for
inclusion in an HTML email, so we now use it, instead of the raw
Markdown version.  Also, because we're using the version that has
already been converted to HTML, we need to use {{{...}}} to display
it, instead of {{...}} so HTML entities aren't escaped.