Update specs to work with development version of octocat_herder gem
The development version of octocat_herder will retrieve the text for
the diff, and patch of a pull request when generating the hash version
of the pull request.  Because of this, we need to stub out the methods
that retrieve the data so it won't actually make an API request.