Add diff_text to example individual opened template
One comment that has been received is that the over-all diff of the
pull request would be handy to have in the notification email.

While this is not suitable for doing the code review since the
individual commits should be examined to make sure that they each do
what it is that they say they do, the over-all diff is still useful
for getting a feel for what the pull request is doing, and whether it
is sufficiently tested.